+++ srcs/fetch/youtube.py
@@ -1,9 +1,9 @@
 # Turn/IP
 # Claire-lex - 2023
-# Interface to search for protocols-related videos on Youtube
+# Interface to fetch protocols-related videos on Youtube
 # pylint: disable=invalid-name,import-error,too-few-public-methods
 
-"""Search for protocols-related videos on Youtube."""
+"""Fetch protocols-related videos on Youtube."""
 
 from importlib.util import find_spec
 try:
@@ -14,7 +14,7 @@ except ModuleNotFoundError as mnfe:
 # Internal
 from config import GOOGLE_API_KEY, youtube as y
 from db import Protocol
-from . import SearchException
+from . import FetchException
 
 #-----------------------------------------------------------------------------#
 # Constants                                                                   #
@@ -51,7 +51,7 @@ class Video(object):
             self.channelId = raw["snippet"]["channelId"]
             self.channel = y.selected_channels[self.channelId]
         except KeyError:
-            raise SearchException(ERR_BADRET)
+            raise FetchException(ERR_BADRET)
 
     def __str__(self):
         return "{0} @ {1} ({2})".format(
@@ -62,7 +62,7 @@ class Video(object):
         If we don't do that the description we have is truncated.
         """
         if not youtube_api:
-            raise SearchException(ERR_YTAPI)
+            raise FetchException(ERR_YTAPI)
         description = youtube_api.videos().list(
             id=self.raw["id"]["videoId"], part="snippet"
         ).execute()
@@ -75,12 +75,12 @@ class Youtube(object):
     def __init__(self):
         # It will raise an exception (not caught this time) if not installed.
         if not find_spec('googleapiclient'):
-            raise SearchException(ERR_GAPI)
+            raise FetchException(ERR_GAPI)
         self.youtube_api = build(y.api_service_name, y.api_version,
                                  developerKey=GOOGLE_API_KEY)
 
     def get_videos(self, protocol: Protocol) -> list:
-        """Get videos about a protocol from selected Youtube channels.
+        """Fetch videos about a protocol from selected Youtube channels.
 
         The list of channels can be modified from config.py.
         """
@@ -98,11 +98,11 @@ class Youtube(object):
                         if item["snippet"]["title"] not in found_titles:
                             try:
                                 video = Video(item, self.youtube_api)
-                            except SearchException:
+                            except FetchException:
                                 pass # Video was invalid
                             else:
                                 candidates.append(video)
                                 found_titles.append(item["snippet"]["title"])
         except HttpError as he:
-            raise SearchException(str(he)) from None
+            raise FetchException(str(he)) from None
         return sorted(set(candidates), key=lambda x: x.year)